+++ src/test/java/seedu/flashnotes/logic/commands/AddCardCommandTest.java
@@ -276,7 +276,7 @@ public class AddCardCommandTest {
         }
 
         @Override
-        public void updateDeckPerformanceScore(Integer reviewScore, String deckName) {
+        public void updateDeckPerformanceScore(Double reviewScore, String deckName) {
             throw new AssertionError("This method should not be called.");
         }
